requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
numpy==1.26.4
faiss-cpu==1.8.0.post1
